fix: fetch delegate before delegate (#4)
getDelegate can in some cases create new Delegation.
With previous order it would result in governane entity from
getGovernance call that is treated as new entity so when we save it
it will result in conflict (because it's now entity as it was already
created by getDelegate).

By changing this order we workaround this issue.

Co-authored-by: Less <bonustrack@users.noreply.github.com>